--- core/core-domain/src/main/kotlin/com/lms/core/domain/problem/domain/exception/InvalidValueException.kt
@@ -1,7 +1,0 @@
-package com.lms.core.domain.problem.domain.exception
-
-import org.apache.coyote.BadRequestException
-
-class InvalidValueException(
-    message: String = "Invalid value exception",
-) : BadRequestException(message)
