+++ core/core-domain/src/main/kotlin/com/lms/core/domain/problem/domain/problem/exception/InvalidValueException.kt
@@ -1,0 +1,7 @@
+package com.lms.core.domain.problem.domain.problem.exception
+
+import com.lms.core.exception.BusinessException
+
+class InvalidValueException(
+    message: String = "Invalid value exception",
+) : BusinessException(message)
