--- core/core-enum/src/main/kotlin/com/lms/enum/exception/InvalidLevelValueException.kt
@@ -1,8 +1,0 @@
-package com.lms.enum.exception
-
-import org.apache.coyote.BadRequestException
-
-class InvalidLevelValueException(
-    levelValue: Int,
-    message: String = "Invalid level value $levelValue"
-) : BadRequestException(message)
